requirements.txt

google-genai
requests


SpeechRecognition
PyAudio
pygame


PyQt5


pywhatkit
pyautogui
pillow


opencv-python
face-recognition
numpy
cmake
dlib